% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessIntervals.R
\name{calculateGCContentByInterval}
\alias{calculateGCContentByInterval}
\title{Calculates GC content by interval}
\usage{
calculateGCContentByInterval(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{preprocessIntervals}}.}
}
\description{
This function was renamed to \code{\link{preprocessIntervals}}.
}
