% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctCoverageBias.R
\name{correctCoverageBias}
\alias{correctCoverageBias}
\title{Correct for GC bias}
\usage{
correctCoverageBias(coverage.file, gc.gene.file, output.file = NULL,
  plot.gc.bias = FALSE, plot.max.density = 50000)
}
\arguments{
\item{coverage.file}{Coverage file or coverage data parsed with the
\code{\link{readCoverageFile}} function.}

\item{gc.gene.file}{File providing GC content for each exon in the coverage
files. First column in format CHR:START-END. Second column GC content (0 to
1).  Third column provides gene symbols, which are optional, but used in
\code{\link{runAbsoluteCN}} to generate gene level calls. This file can be
generated with GATK GCContentByInterval tool or with the
\code{\link{calculateGCContentByInterval}} function.}

\item{output.file}{Optionally, write file with GC corrected coverage. Can be
read with the \code{\link{readCoverageFile}} function.}

\item{plot.gc.bias}{Optionally, plot GC profiles of the pre-normalized and
post-normalized coverage. Provides a quick visual check of coverage bias.}

\item{plot.max.density}{By default, if the number of intervals in the
probe-set is > 50000, uses a kernel density estimate to plot the coverage
distribution. This uses the \code{stat_density} function from the ggplot2
package. Using this parameter, change the threshold at which density
estimation is applied. If the \code{plot.gc.bias} parameter is set as
\code{FALSE}, this will be ignored.}
}
\description{
Takes as input coverage data and a mapping file for GC content. Will then
normalize coverage data for GC-bias.  Optionally plots the pre and
post normalization GC profiles.
}
\examples{

normal.coverage.file <- system.file("extdata", "example_normal.txt", 
    package="PureCN")
gc.gene.file <- system.file("extdata", "example_gc.gene.file.txt", 
    package="PureCN")
coverage <- correctCoverageBias(normal.coverage.file, gc.gene.file)

}
\seealso{
\code{\link{calculateGCContentByInterval}}
}
\author{
Angad Singh, Markus Riester
}
