% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateIntervalWeights.R
\name{createTargetWeights}
\alias{createTargetWeights}
\title{Calculate target weights}
\usage{
createTargetWeights()
}
\description{
This function is defunct, use \code{\link{calculateIntervalWeights}}
instead.
}
\author{
Markus Riester
}
