% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterIntervals.R
\name{filterTargets}
\alias{filterTargets}
\title{Remove low quality targets}
\usage{
filterTargets(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{filterIntervals}}.}
}
\description{
This function is deprecated and was renamed to 
\code{\link{filterIntervals}}.
}
