\name{PureCN-deprecated}
\alias{PureCN-deprecated}
\title{Deprecated functions in package \sQuote{PureCN}}

\description{
  These functions are provided for compatibility with older versions
  of \sQuote{PureCN} only, and will be defunct at the next release.
}

\details{
  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{
    \item{createExonWeightFile: \code{\link{createTargetWeights}}}
  }
}
