\name{purecn.DNAcopy.bdry}
\docType{data}
\alias{purecn.DNAcopy.bdry}
\title{DNAcopy boundary data}
\description{
This provides the output of the \code{DNAcopy::getbdry} call using \code{\link{segmentationCBS}}
default parameters.
}
\usage{data(purecn.DNAcopy.bdry)}
\value{Output of the \code{DNAcopy::getbdry} call.}
\keyword{datasets}
