\name{purecn.example.output}
\docType{data}
\alias{purecn.example.output}
\title{Example output}
\description{
This provides the output of the \code{\link{runAbsoluteCN}} call used in the 
vignette and examples.
}
\usage{data(purecn.example.output)}
\value{Output of the \code{\link{runAbsoluteCN}} call used in the vignette.}
\keyword{datasets}
