% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCoverageFile.R
\name{readCoverageFile}
\alias{readCoverageFile}
\title{Read coverage file}
\usage{
readCoverageFile(file, format)
}
\arguments{
\item{file}{Target coverage file.}

\item{format}{File format. If missing, derived from the file 
extension. Currently only GATK DepthofCoverage format supported.}
}
\value{
A \code{data.frame} with the parsed coverage information.
}
\description{
Read coverage file produced by The Genome Analysis Toolkit or by
\code{\link{calculateBamCoverageByInterval}}.
}
\examples{

tumor.coverage.file <- system.file("extdata", "example_tumor.txt", 
    package="PureCN")
coverage <- readCoverageFile(tumor.coverage.file)

}
\seealso{
\code{\link{calculateBamCoverageByInterval}}
}
\author{
Markus Riester
}
