% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCoverageFile.R
\name{readCoverageGatk}
\alias{readCoverageGatk}
\title{Read GATK coverage files}
\usage{
readCoverageGatk(file)
}
\arguments{
\item{file}{Exon coverage file as produced by GATK.}
}
\value{
A \code{data.frame} with the parsed coverage information.
}
\description{
Read coverage file produced by The Genome Analysis Toolkit or by
\code{\link{calculateBamCoverageByInterval}}. This function
is deprecated. Please use \code{\link{readCoverageFile}} instead.
}
\seealso{
\code{\link{calculateBamCoverageByInterval}}
\code{\link{readCoverageFile}}
}
\author{
Markus Riester
}
