% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curateResults.R
\name{autoCurateResults}
\alias{autoCurateResults}
\title{Heuristics to find the best purity/ploidy solution.}
\usage{
autoCurateResults()
}
\description{
This function is defunct.
}
\author{
Markus Riester
}
