% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curateResults.R
\name{autoCurateResults}
\alias{autoCurateResults}
\title{Heuristics to find the best purity/ploidy solution.}
\usage{
autoCurateResults(res, bootstrap = TRUE, bootstrap.n = 500,
  verbose = TRUE)
}
\arguments{
\item{res}{Return object of the \code{\link{runAbsoluteCN}} function.}

\item{bootstrap}{Try to reduce the number of local optima by using the
bootstrapResults function.}

\item{bootstrap.n}{Number of bootstrap replicates.}

\item{verbose}{Verbose output.}
}
\value{
The provided \code{\link{runAbsoluteCN}} return object with unlikely
purity and ploidy solutions filtered out.
}
\description{
This function is deprecated and will be made defunct in next release.
}
\details{
This implements a workflow with various heuristics, with the goal of
identifying correct purity/ploidy solutions in difficult samples. This is
mainly for automated copy number calling. This function may evolve over time
and might produce different rankings after PureCN updates.
}
\seealso{
\code{\link{runAbsoluteCN}}
}
\author{
Markus Riester
}
