% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateIntervalWeights.R
\name{calculateIntervalWeights}
\alias{calculateIntervalWeights}
\title{Calculate interval weights}
\usage{
calculateIntervalWeights()
}
\description{
This function is now automatically called by \code{\link{createNormalDatabase}}
and is thus defunct.
}
\author{
Markus Riester
}
