% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNormalDatabase.R
\name{createExonWeightFile}
\alias{createExonWeightFile}
\title{Calculate exon weights}
\usage{
createExonWeightFile()
}
\description{
This function is defunct. Please use \code{\link{createTargetWeights}}
instead.
}
\author{
Markus Riester
}
