% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSNPBlacklist.R
\name{createSNPBlacklist}
\alias{createSNPBlacklist}
\title{Create SNP black list}
\usage{
createSNPBlacklist()
}
\description{
This function is defunct If a pool of normal VCF is available, please
use the \code{normal.panel.vcf.file} argument of the
\code{\link{setMappingBiasVcf}} function.
}
\author{
Markus Riester
}
