% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateIntervalWeights.R
\name{createTargetWeights}
\alias{createTargetWeights}
\title{Calculate target weights}
\usage{
createTargetWeights(normal.coverage.files, target.weight.file, plot = FALSE)
}
\arguments{
\item{normal.coverage.files}{A set of normal coverage samples
to estimate target log-ratio standard deviations.}

\item{target.weight.file}{Output filename.}

\item{plot}{Diagnostics plot, useful to tune parameters.}
}
\value{
A \code{data.frame} with target weights.
}
\description{
This function is deprecated, use \code{\link{calculateIntervalWeights}}
instead.
}
\author{
Markus Riester
}
