% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNormalDatabase.R
\name{createTargetWeights}
\alias{createTargetWeights}
\title{Calculate target weights}
\usage{
createTargetWeights(normal.coverage.files, target.weight.file, plot = FALSE)
}
\arguments{
\item{normal.coverage.files}{A set of normal coverage samples
to estimate target log-ratio standard deviations.}

\item{target.weight.file}{Output filename.}

\item{plot}{Diagnostics plot, useful to tune parameters.}
}
\value{
A \code{data.frame} with target weights.
}
\description{
Creates a target weight file useful for segmentation. Requires a set of 
coverage files from normal samples. Target weights will be
set proportional to the inverse of coverage standard deviation across all
normals. Targets with high variance in coverage in the pool of normals are
thus down-weighted.
}
\examples{

target.weight.file <- "target_weights.txt"
normal.coverage.file <- system.file("extdata", "example_normal.txt", 
    package="PureCN")
normal2.coverage.file <- system.file("extdata", "example_normal2.txt", 
    package="PureCN")
normal.coverage.files <- c(normal.coverage.file, normal2.coverage.file)

createTargetWeights(normal.coverage.files, target.weight.file)

}
\author{
Markus Riester
}
