% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterTargets.R
\name{filterTargets}
\alias{filterTargets}
\title{Remove low quality targets}
\usage{
filterTargets(normal, tumor, log.ratio, seg.file, filter.lowhigh.gc = 0.001,
  min.coverage = 15, min.targeted.base = 5, normalDB = NULL)
}
\arguments{
\item{normal}{Coverage data for normal sample.}

\item{tumor}{Coverage data for tumor sample.}

\item{log.ratio}{Copy number log-ratios, one for each target or interval in
coverage file.}

\item{seg.file}{If not \code{NULL}, then do not filter targets, because data
is already segmented via the provided segmentation file.}

\item{filter.lowhigh.gc}{Quantile q (defines lower q and upper 1-q) for
removing targets with outlier GC profile. Assuming that GC correction might
not have been worked on those. Requires \code{interval.file}.}

\item{min.coverage}{Minimum coverage in both normal and tumor. Targets with
lower coverage are ignored. If a \code{normalDB} is provided, then this
database already provides information about low quality targets and the
\code{min.coverage} is set to \code{min.coverage/10000}.}

\item{min.targeted.base}{Exclude intervals with targeted base (size in bp)
smaller than this cutoff. This is useful when the same interval file was
used to calculate GC content. For such small targets, the GC content is
likely very different from the true GC content of the probes.}

\item{normalDB}{Normal database, created with
\code{\link{createNormalDatabase}}.}
}
\value{
\code{logical(length(log.ratio))} specifying which targets should be
used in segmentation.
}
\description{
This function determines which intervals in the coverage files should be
included or excluded in the segmentation. It is called via the
\code{fun.filterTargets} argument of \code{\link{runAbsoluteCN}}. The
arguments are passed via \code{args.filterTargets}.
}
\examples{

normal.coverage.file <- system.file("extdata", "example_normal.txt", 
    package="PureCN")
normal2.coverage.file <- system.file("extdata", "example_normal2.txt", 
    package="PureCN")
normal.coverage.files <- c(normal.coverage.file, normal2.coverage.file)
normalDB <- createNormalDatabase(normal.coverage.files)

tumor.coverage.file <- system.file("extdata", "example_tumor.txt", 
    package="PureCN")
vcf.file <- system.file("extdata", "example_vcf.vcf", 
    package="PureCN")
interval.file <- system.file("extdata", "example_intervals.txt", 
    package="PureCN")

# The max.candidate.solutions, max.ploidy and test.purity parameters are set to
# non-default values to speed-up this example.  This is not a good idea for real
# samples.
ret <-runAbsoluteCN(normal.coverage.file=normal.coverage.file,
    tumor.coverage.file=tumor.coverage.file, genome="hg19", vcf.file=vcf.file,
    sampleid="Sample1", interval.file=interval.file, normalDB=normalDB,
    args.filterTargets=list(min.targeted.base=10), max.ploidy=4, 
    test.purity=seq(0.3,0.7,by=0.05), max.candidate.solutions=1)

}
\author{
Markus Riester
}
