% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBestNormal.R
\name{findBestNormal}
\alias{findBestNormal}
\title{Find best normal sample in database}
\usage{
findBestNormal()
}
\description{
Function to find the best matching normal for a provided tumor sample.
This function is defunct and replaced by
\code{\link{calculateTangentNormal}}.
}
