\name{PureCN-defunct}
\alias{PureCN-defunct}
\title{Defunct functions in package \sQuote{PureCN}}

\description{
  These functions are defunct and no longer available.
}

\details{
  The following functions are defunct; use
  the replacement indicated below:
  \itemize{
    \item{autoCurateResults: no replacement}
    \item{createExonWeightFile: \code{\link{createTargetWeights}}}
    \item{createSNPBlacklist: \code{\link{setMappingBiasVcf}}}
    \item{getDiploid: no replacement}
  }
}
