% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCoverageFile.R
\name{readCoverageGatk}
\alias{readCoverageGatk}
\title{Read GATK coverage files}
\usage{
readCoverageGatk()
}
\description{
This function is deprecated. Please use 
\code{\link{readCoverageFile}} instead.
}
\seealso{
\code{\link{calculateBamCoverageByInterval}}
\code{\link{readCoverageFile}}
}
\author{
Markus Riester
}
